+++ src/rules.vue
@@ -657,11 +657,26 @@ export default {
     },
 
     deleteRule(uuid) {
-      axios.delete(`config-editor/rules/${uuid}`, this.$root.axiosOptions)
-        .then(() => {
-          this.$bus.$emit(constants.NOTIFY_CHANGE_PENDING, true)
+      this.$bvModal.msgBoxConfirm('Do you really want to delete this rule?', {
+        buttonSize: 'sm',
+        cancelTitle: 'NO',
+        centered: true,
+        okTitle: 'YES',
+        okVariant: 'danger',
+        size: 'sm',
+        title: 'Please Confirm',
+      })
+        .then(val => {
+          if (!val) {
+            return
+          }
+
+          return axios.delete(`config-editor/rules/${uuid}`, this.$root.axiosOptions)
+            .then(() => {
+              this.$bus.$emit(constants.NOTIFY_CHANGE_PENDING, true)
+            })
+            .catch(err => this.$bus.$emit(constants.NOTIFY_FETCH_ERROR, err))
         })
-        .catch(err => this.$bus.$emit(constants.NOTIFY_FETCH_ERROR, err))
     },
 
     editRule(msg) {
